// src/app/core/Services/models/addProduct.interface.ts
// export default interface addProduct {
//     name: string;
//     description: string;
//     price: number;
//     stock: number;
//     image: string;
//     categoryId: number;
// }

export default interface addProduct {
  name: string;
  description: string;
  price: number;
  sku: string;
  stock: number;
  categoryId: number;
  productImgUrl: string;
}